+++ page.php
@@ -22,7 +22,7 @@ get_header(); ?>
 
 				do_action( 'shapla_page_before' );
 
-				get_template_part( 'content', 'page' );
+				get_template_part( 'template-parts/content', 'page' );
 
 				/**
 				 * Functions hooked in to shapla_page_after action