+++ search.php
@@ -9,25 +9,21 @@
 
 get_header(); ?>
 
-    <section id="primary" class="content-area">
+    <div id="primary" class="content-area">
         <main id="main" class="site-main" role="main">
 
-			<?php
-			if ( have_posts() ) :
-				while ( have_posts() ) : the_post();
+			<?php if ( have_posts() ) :
 
-					get_template_part( 'content', 'search' );
-
-				endwhile;
+				get_template_part( 'loop' );
 
 			else :
 
-				get_template_part( 'content', 'none' );
+				get_template_part( 'template-parts/content', 'none' );
 
 			endif; ?>
 
         </main><!-- #main -->
-    </section><!-- #primary -->
+    </div><!-- #primary -->
 
 <?php
 get_sidebar();